expose the api server from the api package
